refactor: test project structure should mirror the structure of the system under test
