+++ tests/ParticipantManager.API.Tests/Functions/ParticipantFunctionsTests.cs
@@ -1,4 +1,3 @@
-
 using System.ComponentModel.DataAnnotations;
 using System.Text.Json;
 using Microsoft.AspNetCore.Http;
@@ -11,7 +10,7 @@ using ParticipantManager.API.Functions;
 using ParticipantManager.API.Models;
 using ParticipantManager.TestUtils;
 
-namespace ParticipantManager.API.Tests;
+namespace ParticipantManager.API.Tests.Functions;
 
 public class ParticipantFunctionsTests
 {
