+++ tests/ParticipantManager.Experience.API.Tests/Functions/PathwayEnrolmentFunctionTests.cs
@@ -10,46 +10,47 @@ using ParticipantManager.Experience.API.Services;
 using ParticipantManager.Shared.Client;
 using ParticipantManager.Shared.DTOs;
 
-namespace ParticipantManager.Experience.API.Tests;
+namespace ParticipantManager.Experience.API.Tests.Functions;
 
-public class ScreeningEligibilityFunctionTests
+public class PathwayEnrolmentFunctionTests
 {
     private readonly Mock<ICrudApiClient> _crudApiClient = new();
-    private readonly ScreeningEligibilityFunction _function;
-    private readonly Mock<ILogger<ScreeningEligibilityFunction>> _loggerMock;
+    private readonly PathwayEnrolmentFunction _function;
+    private readonly Mock<ILogger<PathwayEnrolmentFunction>> _loggerMock;
     private readonly Mock<ITokenService> _mockTokenService = new();
     private readonly Mock<IFeatureFlagClient> _mockFeatureFlagClient = new();
-    private readonly HttpRequestData _request = CreateHttpRequest("");
     private readonly Guid _participantId = Guid.NewGuid();
+    private readonly Guid _enrolmentId = Guid.NewGuid();
+    private readonly HttpRequestData _request = CreateHttpRequest("");
 
-    public ScreeningEligibilityFunctionTests()
+    public PathwayEnrolmentFunctionTests()
     {
-        _loggerMock = new Mock<ILogger<ScreeningEligibilityFunction>>();
-        _crudApiClient.Setup(s => s.GetPathwayEnrolmentsAsync(It.IsAny<Guid>()).Result)
-            .Returns(MockListPathwayEnrolments);
-        _function = new ScreeningEligibilityFunction(_loggerMock.Object, _crudApiClient.Object,
-            _mockTokenService.Object,
+        _loggerMock = new Mock<ILogger<PathwayEnrolmentFunction>>();
+        _crudApiClient.Setup(s => s.GetPathwayEnrolmentByIdAsync(It.IsAny<Guid>(), It.IsAny<Guid>()).Result)
+            .Returns(MockPathwayDetails);
+        _function = new PathwayEnrolmentFunction(_loggerMock.Object, _crudApiClient.Object, _mockTokenService.Object,
             _mockFeatureFlagClient.Object);
-        _mockFeatureFlagClient.Setup(f => f.IsFeatureEnabledForParticipant("mays_mvp", _participantId))
+        _mockFeatureFlagClient.Setup(f => f.IsFeatureEnabledForParticipant(It.IsAny<string>(), It.IsAny<Guid>()))
             .ReturnsAsync(true);
     }
 
     [Fact]
-    public async Task GetScreeningEligibility_InvalidToken_ReturnsUnauthorized()
+    public async Task GetPathwayEnrolmentById_InvalidToken_ReturnsUnauthorized()
     {
         // Arrange
         _mockTokenService.Setup(s => s.ValidateToken(It.IsAny<HttpRequestData>()))
             .ReturnsAsync(AccessTokenResult.Expired());
 
         // Act
-        var response = await _function.GetScreeningEligibility(_request, Guid.NewGuid()) as UnauthorizedResult;
+        var response =
+            await _function.GetPathwayEnrolmentById(_request, _participantId, _enrolmentId) as UnauthorizedResult;
 
         // Assert
         Assert.Equal(StatusCodes.Status401Unauthorized, response?.StatusCode);
     }
 
     [Fact]
-    public async Task GetScreeningEligibility_NoNhsNumber_ReturnsUnauthorized()
+    public async Task GetPathwayEnrolmentById_NoNhsNumber_ReturnsUnauthorized()
     {
         // Arrange
         var claims = new List<Claim>
@@ -66,21 +67,22 @@ public class ScreeningEligibilityFunctionTests
             .ReturnsAsync(AccessTokenResult.Success(principal));
 
         // Act
-        var response = await _function.GetScreeningEligibility(_request, Guid.NewGuid()) as UnauthorizedResult;
+        var response =
+            await _function.GetPathwayEnrolmentById(_request, _participantId, _enrolmentId) as UnauthorizedResult;
 
         // Assert
         Assert.Equal(StatusCodes.Status401Unauthorized, response?.StatusCode);
     }
 
     [Fact]
-    public async Task GetScreeningEligibility_ParticipantIdIsEmpty_ReturnsUnauthorized()
+    public async Task GetPathwayEnrolmentById_PathwayEnrolmentIsNull_ReturnsNotFound()
     {
         // Arrange
         var claims = new List<Claim>
         {
-            new(ClaimTypes.NameIdentifier, "12345"),
+            new(ClaimTypes.NameIdentifier, "1234567890"),
             new(ClaimTypes.Email, "user@example.com"),
-            new("nhs_number", "12345678")
+            new("nhs_number", "1234567890")
         };
 
         var identity = new ClaimsIdentity(claims, "Bearer");
@@ -88,50 +90,26 @@ public class ScreeningEligibilityFunctionTests
 
         _mockTokenService.Setup(s => s.ValidateToken(It.IsAny<HttpRequestData>()))
             .ReturnsAsync(AccessTokenResult.Success(principal));
+        _crudApiClient.Setup(s => s.GetPathwayEnrolmentByIdAsync(_participantId, _enrolmentId))
+            .Returns(Task.FromResult<EnrolledPathwayDetailsDto?>(null));
 
         // Act
-        var response = await _function.GetScreeningEligibility(_request, Guid.Empty) as UnauthorizedResult;
+        var response =
+            await _function.GetPathwayEnrolmentById(_request, _participantId, _enrolmentId) as NotFoundResult;
 
         // Assert
-        Assert.Equal(StatusCodes.Status401Unauthorized, response?.StatusCode);
+        Assert.Equal(StatusCodes.Status404NotFound, response?.StatusCode);
     }
 
     [Fact]
-    public async Task GetScreeningEligibility_PathwayEnrolmentsReturnsEmptyCollection_ReturnsOk()
+    public async Task GetPathwayEnrolmentById_NhsNumberDoesNotMatch_ReturnsForbidden()
     {
         // Arrange
         var claims = new List<Claim>
         {
             new(ClaimTypes.NameIdentifier, "12345"),
             new(ClaimTypes.Email, "user@example.com"),
-            new("nhs_number", "12345678")
-        };
-
-        var identity = new ClaimsIdentity(claims, "Bearer");
-        var principal = new ClaimsPrincipal(identity);
-
-        _mockTokenService.Setup(s => s.ValidateToken(It.IsAny<HttpRequestData>()))
-            .ReturnsAsync(AccessTokenResult.Success(principal));
-        _crudApiClient.Setup(s => s.GetPathwayEnrolmentsAsync(_participantId)).ReturnsAsync([]);
-
-        // Act
-        var response = await _function.GetScreeningEligibility(_request, _participantId) as OkObjectResult;
-
-        // Assert
-        Assert.IsType<OkObjectResult>(response);
-        Assert.Equal(StatusCodes.Status200OK, response.StatusCode);
-        Assert.NotNull(response);
-    }
-
-    [Fact]
-    public async Task GetScreeningEligibility_NhsNumberDoesNotMatch_ReturnsForbidden()
-    {
-        // Arrange
-        var claims = new List<Claim>
-        {
-            new(ClaimTypes.NameIdentifier, "12345"),
-            new(ClaimTypes.Email, "user@example.com"),
-            new("nhs_number", "87654321") // Different NHS number than in mock enrolments
+            new("nhs_number", "0987654321") // Different from the mocked data
         };
 
         var identity = new ClaimsIdentity(claims, "Bearer");
@@ -141,7 +119,8 @@ public class ScreeningEligibilityFunctionTests
             .ReturnsAsync(AccessTokenResult.Success(principal));
 
         // Act
-        var response = await _function.GetScreeningEligibility(_request, _participantId);
+        var response =
+            await _function.GetPathwayEnrolmentById(_request, _participantId, _enrolmentId);
 
         // Assert
         Assert.NotNull(response);
@@ -149,14 +128,14 @@ public class ScreeningEligibilityFunctionTests
     }
 
     [Fact]
-    public async Task GetScreeningEligibility_FeatureIsDisabled_ReturnsForbidden()
+    public async Task GetPathwayEnrolmentById_FeatureToggleIsDisabled_ReturnsForbidden()
     {
         // Arrange
         var claims = new List<Claim>
         {
-            new(ClaimTypes.NameIdentifier, "12345678"),
+            new(ClaimTypes.NameIdentifier, "1234567890"),
             new(ClaimTypes.Email, "user@example.com"),
-            new("nhs_number", "12345678")
+            new("nhs_number", "1234567890")
         };
 
         var identity = new ClaimsIdentity(claims, "Bearer");
@@ -168,7 +147,7 @@ public class ScreeningEligibilityFunctionTests
             .ReturnsAsync(false);
 
         // Act
-        var response = await _function.GetScreeningEligibility(_request, _participantId);
+        var response = await _function.GetPathwayEnrolmentById(_request, _participantId, _enrolmentId);
 
         // Assert
         Assert.NotNull(response);
@@ -176,16 +155,15 @@ public class ScreeningEligibilityFunctionTests
     }
 
     [Fact]
-    public async Task GetScreeningEligibility_ValidToken_ReturnsOk()
+    public async Task GetPathwayEnrolmentById_ValidToken_ReturnsOk()
     {
         // Arrange
         var claims = new List<Claim>
         {
-            new(ClaimTypes.NameIdentifier, "12345678"),
+            new(ClaimTypes.NameIdentifier, "1234567890"),
             new(ClaimTypes.Email, "user@example.com"),
-            new("nhs_number", "12345678")
+            new("nhs_number", "1234567890")
         };
-
         var identity = new ClaimsIdentity(claims, "Bearer");
         var principal = new ClaimsPrincipal(identity);
 
@@ -193,73 +171,67 @@ public class ScreeningEligibilityFunctionTests
             .ReturnsAsync(AccessTokenResult.Success(principal));
 
         // Act
-        var response = await _function.GetScreeningEligibility(_request, _participantId) as OkObjectResult;
+        var response =
+            await _function.GetPathwayEnrolmentById(_request, _participantId, _enrolmentId) as OkObjectResult;
 
         // Assert
         Assert.Equal(StatusCodes.Status200OK, response?.StatusCode);
-        var enrollments = response?.Value as List<PathwayEnrolmentDto>;
-        Assert.NotNull(enrollments);
-        Assert.Equal(2, enrollments.Count);
+        var pathwayEnrolmentDto = response?.Value as EnrolledPathwayDetailsDto;
+        Assert.Equal(pathwayEnrolmentDto?.ScreeningName, pathwayEnrolmentDto?.ScreeningName);
     }
 
     [Fact]
-    public async Task GetScreeningEligibility_ExceptionIsThrown_ReturnsBadRequest()
+    public async Task GetPathwayEnrolmentById_ExceptionIsThrown_ReturnsBadRequest()
     {
         // Arrange
         var claims = new List<Claim>
         {
-            new(ClaimTypes.NameIdentifier, "12345678"),
+            new(ClaimTypes.NameIdentifier, "1234567890"),
             new(ClaimTypes.Email, "user@example.com"),
-            new("nhs_number", "12345678")
+            new("nhs_number", "1234567890")
         };
-
         var identity = new ClaimsIdentity(claims, "Bearer");
         var principal = new ClaimsPrincipal(identity);
 
         _mockTokenService.Setup(s => s.ValidateToken(It.IsAny<HttpRequestData>()))
             .ReturnsAsync(AccessTokenResult.Success(principal));
-        _crudApiClient.Setup(s => s.GetPathwayEnrolmentsAsync(_participantId))
+        _crudApiClient.Setup(s => s.GetPathwayEnrolmentByIdAsync(_participantId, _enrolmentId))
             .ThrowsAsync(new Exception("Test exception message"));
 
         // Act
-        var response = await _function.GetScreeningEligibility(_request, _participantId) as BadRequestObjectResult;
+        var response =
+            await _function.GetPathwayEnrolmentById(_request, _participantId, _enrolmentId) as BadRequestObjectResult;
 
         // Assert
         Assert.Equal(StatusCodes.Status400BadRequest, response?.StatusCode);
         Assert.Equal("Test exception message", response?.Value);
     }
 
+    // ✅ Helper Method to Create Mock HTTP Request
     private static HttpRequestData CreateHttpRequest(string? authHeader)
     {
         var context = new Mock<FunctionContext>();
         var request = new Mock<HttpRequestData>(MockBehavior.Strict, context.Object);
         var headers = new HttpHeadersCollection(new List<KeyValuePair<string, string>>());
         if (!string.IsNullOrEmpty(authHeader)) headers.Add("Authorization", $"{authHeader}");
+
         request.Setup(r => r.Headers).Returns(headers);
         return request.Object;
     }
 
-    private List<PathwayEnrolmentDto> MockListPathwayEnrolments()
+    private EnrolledPathwayDetailsDto MockPathwayDetails()
     {
-        return new List<PathwayEnrolmentDto>
+        return new EnrolledPathwayDetailsDto
         {
-            new()
+            EnrolmentId = Guid.NewGuid(),
+            ScreeningName = "Breast Screening",
+            Status = "Active",
+            EnrolmentDate = DateTime.Now,
+            PathwayTypeName = "Breast Screening Regular",
+            NextActionDate = DateTime.Now,
+            Participant = new ParticipantDto
             {
-                EnrolmentId = "123",
-                ScreeningName = "BreastScreening",
-                Participant = new ParticipantDto
-                {
-                    NhsNumber = "12345678",
-                }
-            },
-            new()
-            {
-                EnrolmentId = "1234",
-                ScreeningName = "BowelScreening",
-                Participant = new ParticipantDto
-                {
-                    NhsNumber = "12345678",
-                }
+                NhsNumber = "1234567890"
             }
         };
     }
