+++ tests/ParticipantManager.Experience.API.Tests/Functions/ScreeningEligibilityFunctionTests.cs
@@ -10,47 +10,46 @@ using ParticipantManager.Experience.API.Services;
 using ParticipantManager.Shared.Client;
 using ParticipantManager.Shared.DTOs;
 
-namespace ParticipantManager.Experience.API.Tests;
+namespace ParticipantManager.Experience.API.Tests.Functions;
 
-public class PathwayEnrolmentFunctionTests
+public class ScreeningEligibilityFunctionTests
 {
     private readonly Mock<ICrudApiClient> _crudApiClient = new();
-    private readonly PathwayEnrolmentFunction _function;
-    private readonly Mock<ILogger<PathwayEnrolmentFunction>> _loggerMock;
+    private readonly ScreeningEligibilityFunction _function;
+    private readonly Mock<ILogger<ScreeningEligibilityFunction>> _loggerMock;
     private readonly Mock<ITokenService> _mockTokenService = new();
     private readonly Mock<IFeatureFlagClient> _mockFeatureFlagClient = new();
-    private readonly Guid _participantId = Guid.NewGuid();
-    private readonly Guid _enrolmentId = Guid.NewGuid();
     private readonly HttpRequestData _request = CreateHttpRequest("");
+    private readonly Guid _participantId = Guid.NewGuid();
 
-    public PathwayEnrolmentFunctionTests()
+    public ScreeningEligibilityFunctionTests()
     {
-        _loggerMock = new Mock<ILogger<PathwayEnrolmentFunction>>();
-        _crudApiClient.Setup(s => s.GetPathwayEnrolmentByIdAsync(It.IsAny<Guid>(), It.IsAny<Guid>()).Result)
-            .Returns(MockPathwayDetails);
-        _function = new PathwayEnrolmentFunction(_loggerMock.Object, _crudApiClient.Object, _mockTokenService.Object,
+        _loggerMock = new Mock<ILogger<ScreeningEligibilityFunction>>();
+        _crudApiClient.Setup(s => s.GetPathwayEnrolmentsAsync(It.IsAny<Guid>()).Result)
+            .Returns(MockListPathwayEnrolments);
+        _function = new ScreeningEligibilityFunction(_loggerMock.Object, _crudApiClient.Object,
+            _mockTokenService.Object,
             _mockFeatureFlagClient.Object);
-        _mockFeatureFlagClient.Setup(f => f.IsFeatureEnabledForParticipant(It.IsAny<string>(), It.IsAny<Guid>()))
+        _mockFeatureFlagClient.Setup(f => f.IsFeatureEnabledForParticipant("mays_mvp", _participantId))
             .ReturnsAsync(true);
     }
 
     [Fact]
-    public async Task GetPathwayEnrolmentById_InvalidToken_ReturnsUnauthorized()
+    public async Task GetScreeningEligibility_InvalidToken_ReturnsUnauthorized()
     {
         // Arrange
         _mockTokenService.Setup(s => s.ValidateToken(It.IsAny<HttpRequestData>()))
             .ReturnsAsync(AccessTokenResult.Expired());
 
         // Act
-        var response =
-            await _function.GetPathwayEnrolmentById(_request, _participantId, _enrolmentId) as UnauthorizedResult;
+        var response = await _function.GetScreeningEligibility(_request, Guid.NewGuid()) as UnauthorizedResult;
 
         // Assert
         Assert.Equal(StatusCodes.Status401Unauthorized, response?.StatusCode);
     }
 
     [Fact]
-    public async Task GetPathwayEnrolmentById_NoNhsNumber_ReturnsUnauthorized()
+    public async Task GetScreeningEligibility_NoNhsNumber_ReturnsUnauthorized()
     {
         // Arrange
         var claims = new List<Claim>
@@ -67,49 +66,21 @@ public class PathwayEnrolmentFunctionTests
             .ReturnsAsync(AccessTokenResult.Success(principal));
 
         // Act
-        var response =
-            await _function.GetPathwayEnrolmentById(_request, _participantId, _enrolmentId) as UnauthorizedResult;
+        var response = await _function.GetScreeningEligibility(_request, Guid.NewGuid()) as UnauthorizedResult;
 
         // Assert
         Assert.Equal(StatusCodes.Status401Unauthorized, response?.StatusCode);
     }
 
     [Fact]
-    public async Task GetPathwayEnrolmentById_PathwayEnrolmentIsNull_ReturnsNotFound()
-    {
-        // Arrange
-        var claims = new List<Claim>
-        {
-            new(ClaimTypes.NameIdentifier, "1234567890"),
-            new(ClaimTypes.Email, "user@example.com"),
-            new("nhs_number", "1234567890")
-        };
-
-        var identity = new ClaimsIdentity(claims, "Bearer");
-        var principal = new ClaimsPrincipal(identity);
-
-        _mockTokenService.Setup(s => s.ValidateToken(It.IsAny<HttpRequestData>()))
-            .ReturnsAsync(AccessTokenResult.Success(principal));
-        _crudApiClient.Setup(s => s.GetPathwayEnrolmentByIdAsync(_participantId, _enrolmentId))
-            .Returns(Task.FromResult<EnrolledPathwayDetailsDto?>(null));
-
-        // Act
-        var response =
-            await _function.GetPathwayEnrolmentById(_request, _participantId, _enrolmentId) as NotFoundResult;
-
-        // Assert
-        Assert.Equal(StatusCodes.Status404NotFound, response?.StatusCode);
-    }
-
-    [Fact]
-    public async Task GetPathwayEnrolmentById_NhsNumberDoesNotMatch_ReturnsForbidden()
+    public async Task GetScreeningEligibility_ParticipantIdIsEmpty_ReturnsUnauthorized()
     {
         // Arrange
         var claims = new List<Claim>
         {
             new(ClaimTypes.NameIdentifier, "12345"),
             new(ClaimTypes.Email, "user@example.com"),
-            new("nhs_number", "0987654321") // Different from the mocked data
+            new("nhs_number", "12345678")
         };
 
         var identity = new ClaimsIdentity(claims, "Bearer");
@@ -119,8 +90,58 @@ public class PathwayEnrolmentFunctionTests
             .ReturnsAsync(AccessTokenResult.Success(principal));
 
         // Act
-        var response =
-            await _function.GetPathwayEnrolmentById(_request, _participantId, _enrolmentId);
+        var response = await _function.GetScreeningEligibility(_request, Guid.Empty) as UnauthorizedResult;
+
+        // Assert
+        Assert.Equal(StatusCodes.Status401Unauthorized, response?.StatusCode);
+    }
+
+    [Fact]
+    public async Task GetScreeningEligibility_PathwayEnrolmentsReturnsEmptyCollection_ReturnsOk()
+    {
+        // Arrange
+        var claims = new List<Claim>
+        {
+            new(ClaimTypes.NameIdentifier, "12345"),
+            new(ClaimTypes.Email, "user@example.com"),
+            new("nhs_number", "12345678")
+        };
+
+        var identity = new ClaimsIdentity(claims, "Bearer");
+        var principal = new ClaimsPrincipal(identity);
+
+        _mockTokenService.Setup(s => s.ValidateToken(It.IsAny<HttpRequestData>()))
+            .ReturnsAsync(AccessTokenResult.Success(principal));
+        _crudApiClient.Setup(s => s.GetPathwayEnrolmentsAsync(_participantId)).ReturnsAsync([]);
+
+        // Act
+        var response = await _function.GetScreeningEligibility(_request, _participantId) as OkObjectResult;
+
+        // Assert
+        Assert.IsType<OkObjectResult>(response);
+        Assert.Equal(StatusCodes.Status200OK, response.StatusCode);
+        Assert.NotNull(response);
+    }
+
+    [Fact]
+    public async Task GetScreeningEligibility_NhsNumberDoesNotMatch_ReturnsForbidden()
+    {
+        // Arrange
+        var claims = new List<Claim>
+        {
+            new(ClaimTypes.NameIdentifier, "12345"),
+            new(ClaimTypes.Email, "user@example.com"),
+            new("nhs_number", "87654321") // Different NHS number than in mock enrolments
+        };
+
+        var identity = new ClaimsIdentity(claims, "Bearer");
+        var principal = new ClaimsPrincipal(identity);
+
+        _mockTokenService.Setup(s => s.ValidateToken(It.IsAny<HttpRequestData>()))
+            .ReturnsAsync(AccessTokenResult.Success(principal));
+
+        // Act
+        var response = await _function.GetScreeningEligibility(_request, _participantId);
 
         // Assert
         Assert.NotNull(response);
@@ -128,14 +149,14 @@ public class PathwayEnrolmentFunctionTests
     }
 
     [Fact]
-    public async Task GetPathwayEnrolmentById_FeatureToggleIsDisabled_ReturnsForbidden()
+    public async Task GetScreeningEligibility_FeatureIsDisabled_ReturnsForbidden()
     {
         // Arrange
         var claims = new List<Claim>
         {
-            new(ClaimTypes.NameIdentifier, "1234567890"),
+            new(ClaimTypes.NameIdentifier, "12345678"),
             new(ClaimTypes.Email, "user@example.com"),
-            new("nhs_number", "1234567890")
+            new("nhs_number", "12345678")
         };
 
         var identity = new ClaimsIdentity(claims, "Bearer");
@@ -147,7 +168,7 @@ public class PathwayEnrolmentFunctionTests
             .ReturnsAsync(false);
 
         // Act
-        var response = await _function.GetPathwayEnrolmentById(_request, _participantId, _enrolmentId);
+        var response = await _function.GetScreeningEligibility(_request, _participantId);
 
         // Assert
         Assert.NotNull(response);
@@ -155,15 +176,16 @@ public class PathwayEnrolmentFunctionTests
     }
 
     [Fact]
-    public async Task GetPathwayEnrolmentById_ValidToken_ReturnsOk()
+    public async Task GetScreeningEligibility_ValidToken_ReturnsOk()
     {
         // Arrange
         var claims = new List<Claim>
         {
-            new(ClaimTypes.NameIdentifier, "1234567890"),
+            new(ClaimTypes.NameIdentifier, "12345678"),
             new(ClaimTypes.Email, "user@example.com"),
-            new("nhs_number", "1234567890")
+            new("nhs_number", "12345678")
         };
+
         var identity = new ClaimsIdentity(claims, "Bearer");
         var principal = new ClaimsPrincipal(identity);
 
@@ -171,67 +193,73 @@ public class PathwayEnrolmentFunctionTests
             .ReturnsAsync(AccessTokenResult.Success(principal));
 
         // Act
-        var response =
-            await _function.GetPathwayEnrolmentById(_request, _participantId, _enrolmentId) as OkObjectResult;
+        var response = await _function.GetScreeningEligibility(_request, _participantId) as OkObjectResult;
 
         // Assert
         Assert.Equal(StatusCodes.Status200OK, response?.StatusCode);
-        var pathwayEnrolmentDto = response?.Value as EnrolledPathwayDetailsDto;
-        Assert.Equal(pathwayEnrolmentDto?.ScreeningName, pathwayEnrolmentDto?.ScreeningName);
+        var enrollments = response?.Value as List<PathwayEnrolmentDto>;
+        Assert.NotNull(enrollments);
+        Assert.Equal(2, enrollments.Count);
     }
 
     [Fact]
-    public async Task GetPathwayEnrolmentById_ExceptionIsThrown_ReturnsBadRequest()
+    public async Task GetScreeningEligibility_ExceptionIsThrown_ReturnsBadRequest()
     {
         // Arrange
         var claims = new List<Claim>
         {
-            new(ClaimTypes.NameIdentifier, "1234567890"),
+            new(ClaimTypes.NameIdentifier, "12345678"),
             new(ClaimTypes.Email, "user@example.com"),
-            new("nhs_number", "1234567890")
+            new("nhs_number", "12345678")
         };
+
         var identity = new ClaimsIdentity(claims, "Bearer");
         var principal = new ClaimsPrincipal(identity);
 
         _mockTokenService.Setup(s => s.ValidateToken(It.IsAny<HttpRequestData>()))
             .ReturnsAsync(AccessTokenResult.Success(principal));
-        _crudApiClient.Setup(s => s.GetPathwayEnrolmentByIdAsync(_participantId, _enrolmentId))
+        _crudApiClient.Setup(s => s.GetPathwayEnrolmentsAsync(_participantId))
             .ThrowsAsync(new Exception("Test exception message"));
 
         // Act
-        var response =
-            await _function.GetPathwayEnrolmentById(_request, _participantId, _enrolmentId) as BadRequestObjectResult;
+        var response = await _function.GetScreeningEligibility(_request, _participantId) as BadRequestObjectResult;
 
         // Assert
         Assert.Equal(StatusCodes.Status400BadRequest, response?.StatusCode);
         Assert.Equal("Test exception message", response?.Value);
     }
 
-    // ✅ Helper Method to Create Mock HTTP Request
     private static HttpRequestData CreateHttpRequest(string? authHeader)
     {
         var context = new Mock<FunctionContext>();
         var request = new Mock<HttpRequestData>(MockBehavior.Strict, context.Object);
         var headers = new HttpHeadersCollection(new List<KeyValuePair<string, string>>());
         if (!string.IsNullOrEmpty(authHeader)) headers.Add("Authorization", $"{authHeader}");
-
         request.Setup(r => r.Headers).Returns(headers);
         return request.Object;
     }
 
-    private EnrolledPathwayDetailsDto MockPathwayDetails()
+    private List<PathwayEnrolmentDto> MockListPathwayEnrolments()
     {
-        return new EnrolledPathwayDetailsDto
+        return new List<PathwayEnrolmentDto>
         {
-            EnrolmentId = Guid.NewGuid(),
-            ScreeningName = "Breast Screening",
-            Status = "Active",
-            EnrolmentDate = DateTime.Now,
-            PathwayTypeName = "Breast Screening Regular",
-            NextActionDate = DateTime.Now,
-            Participant = new ParticipantDto
+            new()
             {
-                NhsNumber = "1234567890"
+                EnrolmentId = "123",
+                ScreeningName = "BreastScreening",
+                Participant = new ParticipantDto
+                {
+                    NhsNumber = "12345678",
+                }
+            },
+            new()
+            {
+                EnrolmentId = "1234",
+                ScreeningName = "BowelScreening",
+                Participant = new ParticipantDto
+                {
+                    NhsNumber = "12345678",
+                }
             }
         };
     }
